Small updates and tweaks
- Updated the default language file again with a couple tweaks.
- Attempted a better fix for isDestroyed. Seemed to be working fine before, but I still felt like it wasn't quite right.
- A couple of other small tweaks here and there.

Jukeboxes still can't be powered in the same way they were before due to changes in 1.20 that made jukeboxes always output redstone signals. I'll try to fix that at some point, but in the meantime using signs is probably the best way to do things if you want playlists to be triggered via redstone.

This should only be used in 1.21+ by 1.21+ clients. Use of ViaBackwards may result in broken features. In my testing music started with the "play" commands can't be stopped once the menu is closed on a 1.20.4 client connecting to 1 1.21.1 server, but a 1.21.1 client can stop the music just fine. It is what it is. *shrug*